+++ src/fem/block/element/linear/surface/triangle/mod.rs
@@ -4,17 +4,18 @@ mod test;
 use super::*;
 
 const G: usize = 1;
-const M: usize = 3;
-const N: usize = 4;
-const O: usize = 4;
+const M: usize = 2;
+const N: usize = 3;
+const O: usize = 3;
 
-pub struct Tetrahedron<C>
+pub struct Triangle<C>
 {
     constitutive_models: [C; G],
-    gradient_vectors: GradientVectors<O>
+    gradient_vectors: GradientVectors<O>,
+    reference_normal: ReferenceNormal
 }
 
-impl<'a, C> FiniteElement<'a, C, G, N> for Tetrahedron<C>
+impl<'a, C> FiniteElement<'a, C, G, N> for Triangle<C>
 where
     C: Constitutive<'a>
 {
@@ -31,27 +32,34 @@ where
         Self
         {
             constitutive_models: std::array::from_fn(|_| <C>::new(constitutive_model_parameters)),
-            gradient_vectors: Self::calculate_gradient_vectors(&reference_nodal_coordinates)
+            gradient_vectors: Self::calculate_gradient_vectors(&reference_nodal_coordinates),
+            reference_normal: Self::calculate_reference_normal(&Self::calculate_reference_dual_basis_vectors(&reference_nodal_coordinates))
         }
     }
 }
 
-impl<'a, C> LinearElement<'a, C, G, M, N, O> for Tetrahedron<C>
+impl<'a, C> LinearElement<'a, C, G, M, N, O> for Triangle<C>
 where
     C: Constitutive<'a>
 {
+    fn calculate_deformation_gradient(&self, nodal_coordinates: &NodalCoordinates<N>) -> DeformationGradient
+    {
+        self.calculate_deformation_gradient_linear_surface_element(nodal_coordinates)
+    }
+    fn calculate_deformation_gradient_rate(&self, nodal_coordinates: &NodalCoordinates<N>, nodal_velocities: &NodalVelocities<N>) -> DeformationGradientRate
+    {
+        self.calculate_deformation_gradient_rate_linear_surface_element(nodal_coordinates, nodal_velocities)
+    }
     fn calculate_gradient_vectors(reference_nodal_coordinates: &ReferenceNodalCoordinates<N>) -> GradientVectors<O>
     {
-        let standard_gradient_operator = Self::calculate_standard_gradient_operator();
-        (reference_nodal_coordinates * &standard_gradient_operator).inverse_transpose() * standard_gradient_operator
+        Self::calculate_gradient_vectors_linear_surface_element(reference_nodal_coordinates)
     }
     fn calculate_standard_gradient_operator() -> StandardGradientOperator<M, O>
     {
         StandardGradientOperator::new([
-            [-1.0, -1.0, -1.0],
-            [ 1.0,  0.0,  0.0],
-            [ 0.0,  1.0,  0.0],
-            [ 0.0,  0.0,  1.0]
+            [-1.0, -1.0],
+            [ 1.0,  0.0],
+            [ 0.0,  1.0]
         ])
     }
     fn get_gradient_vectors(&self) -> &GradientVectors<O>
@@ -60,7 +68,17 @@ where
     }
 }
 
-impl<'a, C> ElasticFiniteElement<'a, C, G, N> for Tetrahedron<C>
+impl<'a, C> LinearSurfaceElement<'a, C, G, M, N, O> for Triangle<C>
+where
+    C: Constitutive<'a>
+{
+    fn get_reference_normal(&self) -> &ReferenceNormal
+    {
+        &self.reference_normal
+    }
+}
+
+impl<'a, C> ElasticFiniteElement<'a, C, G, N> for Triangle<C>
 where
     C: Elastic<'a>
 {
@@ -74,12 +92,12 @@ where
     }
 }
 
-impl<'a, C> ElasticLinearElement<'a, C, G, M, N, O> for Tetrahedron<C>
+impl<'a, C> ElasticLinearElement<'a, C, G, M, N, O> for Triangle<C>
 where
     C: Elastic<'a>
 {}
 
-impl<'a, C> HyperelasticFiniteElement<'a, C, G, N> for Tetrahedron<C>
+impl<'a, C> HyperelasticFiniteElement<'a, C, G, N> for Triangle<C>
 where
     C: Hyperelastic<'a>
 {
@@ -89,12 +107,12 @@ where
     }
 }
 
-impl<'a, C> HyperelasticLinearElement<'a, C, G, M, N, O> for Tetrahedron<C>
+impl<'a, C> HyperelasticLinearElement<'a, C, G, M, N, O> for Triangle<C>
 where
     C: Hyperelastic<'a>
 {}
 
-impl<'a, C> ViscoelasticFiniteElement<'a, C, G, N> for Tetrahedron<C>
+impl<'a, C> ViscoelasticFiniteElement<'a, C, G, N> for Triangle<C>
 where
     C: Viscoelastic<'a>
 {
@@ -108,12 +126,12 @@ where
     }
 }
 
-impl<'a, C> ViscoelasticLinearElement<'a, C, G, M, N, O> for Tetrahedron<C>
+impl<'a, C> ViscoelasticLinearElement<'a, C, G, M, N, O> for Triangle<C>
 where
     C: Viscoelastic<'a>
 {}
 
-impl<'a, C> ElasticHyperviscousFiniteElement<'a, C, G, N> for Tetrahedron<C>
+impl<'a, C> ElasticHyperviscousFiniteElement<'a, C, G, N> for Triangle<C>
 where
     C: ElasticHyperviscous<'a>
 {
@@ -127,12 +145,12 @@ where
     }
 }
 
-impl<'a, C> ElasticHyperviscousLinearElement<'a, C, G, M, N, O> for Tetrahedron<C>
+impl<'a, C> ElasticHyperviscousLinearElement<'a, C, G, M, N, O> for Triangle<C>
 where
     C: ElasticHyperviscous<'a>
 {}
 
-impl<'a, C> HyperviscoelasticFiniteElement<'a, C, G, N> for Tetrahedron<C>
+impl<'a, C> HyperviscoelasticFiniteElement<'a, C, G, N> for Triangle<C>
 where
     C: Hyperviscoelastic<'a>
 {
@@ -142,7 +160,7 @@ where
     }
 }
 
-impl<'a, C> HyperviscoelasticLinearElement<'a, C, G, M, N, O> for Tetrahedron<C>
+impl<'a, C> HyperviscoelasticLinearElement<'a, C, G, M, N, O> for Triangle<C>
 where
     C: Hyperviscoelastic<'a>
 {}
